Partner List Api
