+++ index.js
@@ -1,0 +1,43 @@
+'use strict';
+
+const express = require('express'),
+    app = express(),
+    cors = require('cors'),
+    port = process.env.PORT || 8080,
+    server = process.env.LVM_PUBLIC_URL || `http://localhost:${port}`;
+
+const repositories = require('./repositories')(server);
+
+const PartnerController = require('./controllers/partner');
+
+const partnerController = new PartnerController(repositories);
+
+// middlewares
+app.use(cors());
+app.use(function (req, res, next) {
+    setTimeout(function () {
+        next();
+    }, 100);
+});
+
+app.use(function (req, res, next) {
+    console.log(`${req.method}: ${req.url}`);
+    next();
+});
+
+
+app.get('/partners', partnerController.list);
+
+
+
+
+
+
+
+module.exports = app;
+
+if (module.parent === null || module.parent.filename.indexOf('server.js') > -1) {
+    app.listen(port, () => {
+        console.log(`server running on port ${port} (http://${server})`);
+    });
+}
