+++ index_old.js
@@ -223,7 +223,6 @@
     if (module.parent === null || module.parent.filename.indexOf('server.js') > -1) {
         app.listen(port, () => {
             console.log(`server running on port ${port} (http://${server})`);
-    })
-        ;
+    });
     }
 })();
