+++ package.json
@@ -4,7 +4,7 @@
   "description": "LAS Rest-Service Mockup",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node server.js"
   },
   "author": "",
   "license": "ISC",
