+++ repositories/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const PartnerRepository = require('./partnerRepository');
+
+module.exports = function (server) {
+    return {
+        partnerRepository: new PartnerRepository(server)
+    };
+};
